pyproject.toml
[build-system]
requires = ['setuptools', 'setuptools-scm']
build-backend = "setuptools.build_meta"

[tool.setuptools_scm]
version_scheme = "post-release"
local_scheme = "node-and-date"

[tools.setuptools.packages.find]
where = ["src"]

[project]
name = "bulkdnsquery"
dynamic = ["version"]
readme = "README.md"
description = "Bulk DNS Query Tool"
license = { text = "MIT" }
requires-python = ">3.9"
dependencies = [
    'xlsxwriter',
    'dnspython '
]

[project.urls]
repository = "https://github.com/pfptcommunity/bulkdnsquery"
#homepage = "https://github.com/pfptcommunity/bulkdnsquery"
#documentation = "https://github.com/pfptcommunity/bulkdnsquery"
#changelog = "https://github.com/pfptcommunity/bulkdnsquery"

[project.scripts]
dnscheck = "bulkdnsquery:main"